Enhance turning test
